requirements.txt
allure-pytest==2.13.2
pytest~=8.2.0
requests~=2.31.0